Correct circular imports.
